FilteringNode, FilteringConnection の定義
